pyproject.toml
[project]
name = "biathlon-app"
version = "1.0.0"
description = "Professional Biathlon Analytics Platform for Coaches - Real IBU Data Analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Biathlon Analytics Team", email = "info@biathlon-analytics.cz"}
]
keywords = [
    "biathlon",
    "sports-analytics",
    "ibu",
    "performance-analysis",
    "coaching",
    "data-science"
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    # Data Processing
    "pandas>=2.1.3",
    "numpy>=1.24.3",
    "scipy>=1.11.4",
    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Biathlon Specific
    "biathlonresults",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "ipython>=8.17.0",
    "jupyter>=1.0.0"
]

ml = [
    "scikit-learn>=1.3.2",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "statsmodels>=0.14.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 100
fix = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
